[TM-978] all changes for providers
