+++ src/components/elements/MapPolygonPanel/MapMenuPanelItem.tsx
@@ -5,8 +5,8 @@ import { DetailedHTMLProps, HTMLAttributes } from "react";
 import Text from "@/components/elements/Text/Text";
 import Icon, { IconNames } from "@/components/extensive/Icon/Icon";
 import ModalConfirm from "@/components/extensive/Modal/ModalConfirm";
+import { useMapAreaContext } from "@/context/mapArea.provider";
 import { useModalContext } from "@/context/modal.provider";
-import { useMapAreaPartner } from "@/context/monitoringPartner.provider";
 
 import Button from "../Button/Button";
 import Menu from "../Menu/Menu";
@@ -39,7 +39,7 @@ const MapMenuPanelItem = ({
 }: MapMenuPanelItemProps) => {
   let imageStatus = `IC_${status.toUpperCase().replace(/-/g, "_")}`;
   const { openModal, closeModal } = useModalContext();
-  const { isMonitoring } = useMapAreaPartner();
+  const { isMonitoring } = useMapAreaContext();
   const t = useT();
   const openFormModalHandlerConfirm = () => {
     openModal(
