+++ src/components/elements/MapSidePanel/MapSidePanel.tsx
@@ -7,7 +7,7 @@ import Text from "@/components/elements/Text/Text";
 import Icon, { IconNames } from "@/components/extensive/Icon/Icon";
 import List from "@/components/extensive/List/List";
 import { STATUSES } from "@/constants/statuses";
-import { useMapAreaPartner } from "@/context/monitoringPartner.provider";
+import { useMapAreaContext } from "@/context/mapArea.provider";
 import {
   fetchDeleteV2TerrafundPolygonUuid,
   fetchGetV2TerrafundGeojsonComplete,
@@ -56,7 +56,7 @@ const MapSidePanel = ({
   const [openMenu, setOpenMenu] = useState(false);
   const [clickedButton, setClickedButton] = useState<string>("");
   const checkboxRefs = useRef<HTMLInputElement[]>([]);
-  const { isMonitoring } = useMapAreaPartner();
+  const { isMonitoring } = useMapAreaContext();
 
   const { map } = mapFunctions;
