+++ src/context/mapArea.provider.tsx
@@ -6,13 +6,24 @@ type MapAreaType = {
   isMonitoring: boolean;
   setIsMonitoring: (value: boolean) => void;
   checkIsMonitoringPartner: (projectUuid: string) => Promise<void>;
+  isUserDrawingEnabled: boolean;
+  toggleUserDrawing: (arg0: boolean) => void;
+  toggleAttribute: (arg0: boolean) => void;
+  openEditNewPolygon: boolean;
 };
 
 const MapAreaContext = createContext<MapAreaType | undefined>(undefined);
 
 export const MapAreaProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
   const [isMonitoring, setIsMonitoring] = useState<boolean>(false);
-
+  const [isUserDrawingEnabled, setIsUserDrawingEnabled] = useState<boolean>(false);
+  const [openEditNewPolygon, setOpenEditNewPolygon] = useState<boolean>(false);
+  const toggleUserDrawing = (isDrawing: boolean) => {
+    setIsUserDrawingEnabled(isDrawing);
+  };
+  const toggleAttribute = (isOpen: boolean) => {
+    setOpenEditNewPolygon(isOpen);
+  };
   const checkIsMonitoringPartner = async (projectUuid: string) => {
     try {
       const isMonitoringPartner: any = await fetchGetV2DashboardViewProjectUuid({
@@ -28,13 +39,17 @@ export const MapAreaProvider: React.FC<{ children: ReactNode }> = ({ children })
   const contextValue: MapAreaType = {
     isMonitoring,
     setIsMonitoring,
-    checkIsMonitoringPartner
+    checkIsMonitoringPartner,
+    isUserDrawingEnabled,
+    toggleUserDrawing,
+    toggleAttribute,
+    openEditNewPolygon
   };
 
   return <MapAreaContext.Provider value={contextValue}>{children}</MapAreaContext.Provider>;
 };
 
-export const useMapAreaPartner = () => {
+export const useMapAreaContext = () => {
   const context = useContext(MapAreaContext);
   if (context === undefined) {
     throw new Error("useMonitoring must be used within a MonitoringPartnerProvider");
